# swagger.yaml
...
swagger: "2.0"
info:
  description: This is a API of SaratovNews
  version: 1.0.0
  title: Simple API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /Ometh322/SararovNews/1.0.0
...
schemes:
- https
- http
paths:
  /newsFeed:
    get:
      tags:
      - newsFeed
      summary: List of All News, Events, Polls in News Feed
      operationId: listNewsFeed
      produces:
      - application/json
      - application/xml
      parameters:
      - name: numberOfObject
        in: query
        description: Number of Outputing Objects
        required: false
        type: integer
      - name: startIndex
        in: query
        description: Index of First Object, that need to be Outputed
        required: false
        type: integer
      responses:
        "200":
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Record'
        "404":
          description: News Feed not found
...
  /polls:
    get:
      tags:
      - polls
      summary: List All Polls
      description: Returns the whole Polls
      operationId: listPolls
      produces:
      - application/json
      - application/xml
      parameters: []
      responses:
        "200":
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Polls'
        "404":
          description: Polls not found
    put:
      tags:
      - polls
      summary: Adds a new Poll post
      operationId: addPoll
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: Poll object that needs to be added
        required: true
        schema:
          $ref: '#/definitions/Polls'
      responses:
        "201":
          description: Successful creating
          schema:
            $ref: '#/definitions/Polls'
        "405":
          description: Invalid Input
  /polls/{id}:
    get:
      tags:
      - polls
      summary: Find Poll by ID
      description: Returns a singlePoll
      operationId: getPollById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of Poll to return
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/Polls'
        "400":
          description: Invalid ID supplied
        "404":
          description: Poll not found
    post:
      tags:
      - polls
      summary: Update poll by ID
      operationId: updatePoll
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of Poll that need to be updated
        required: true
        type: integer
        format: int64
      - in: body
        name: body
        description: Polls object that needs to be updated
        required: true
        schema:
          $ref: '#/definitions/Polls'
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Poll not found
        "405":
          description: Validation exception
    delete:
      tags:
      - polls
      summary: Deletes a Poll post
      operationId: deletePoll
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: Poll ID to delete
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Poll not found
definitions:
...
  Polls:
    allOf:
    - $ref: '#/definitions/Record'
    - properties:
        answerOptions:
          type: array
          items:
            type: string
        selectedAnswer:
          type: integer
          format: int64
  Comments:
    type: object
    properties:
      id:
        type: integer
        format: int64
      creatorId:
        type: integer
        format: int64
      message:
        type: string
      publicationDate:
        type: string
        format: date-time
      numberOfLikes:
        type: integer
        format: int32
  Record:
    type: object
    properties:
      id:
        type: integer
        format: int64
      creatorId:
        type: integer
        format: int64
      title:
        type: string
      publicationDate:
        type: string
        format: date-time
      description:
        type: string
      numberOfComments:
        type: integer
        format: int32
      numberOfLikes:
        type: integer
        format: int32
      comments:
        type: array
        items:
          $ref: '#/definitions/Comments'
      typeOfRecord:
        type: string
        enum:
        - NEWS
        - EVENT
        - POLL
...